chore(scripts): remove package-lock.json, bundle Build-agora-sdk.js
