+++ package.json
@@ -2,7 +2,7 @@
     "name": "root",
     "private": true,
     "scripts": {
-        "preinstall": "npm install minimist@1.2.5 edit-json-file@1.4.1 --save-dev && node ./scripts/Build-agora-sdk.js"
+        "preinstall": "node ./scripts/Build-agora-sdk.js"
     },
     "workspaces": {
         "packages": [
@@ -12,10 +12,6 @@
         "nohoist": [
             "**"
         ]
-    },
-    "devDependencies": {
-        "edit-json-file": "^1.4.1",
-        "minimist": "^1.2.5"
     },
     "agora_electron": {
         "platform": "darwin",
