+++ scripts/Build-agora-sdk.js
@@ -1,29 +1,19 @@
-const minimist = require("minimist");
-const { log, arch, platform, setAgoraBuildInfo } = require("./Utils");
-
-const commandsArgs = minimist(process.argv.slice(2));
-// delete useless keys
-delete commandsArgs._;
-
-// see: https://github.com/AgoraIO/Electron-SDK/wiki/Installation-Configuration-in-package.json
-const defaultOptions = {
-    arch,
-    platform,
-    // agora-electron-sdk currently only supports 7.1.2 pre-compilation, but supports 7.1.14 version of electron
-    electron_version: "7.1.2",
-    msvs_version: "2017",
-    silent: false,
-    debug: false,
-    // No special circumstances, please do not set to false
-    prebuilt: true,
-};
-
-const argv = {
-    ...defaultOptions,
-    ...commandsArgs,
-};
-
-log("current agora parameters");
-log(argv);
-
-setAgoraBuildInfo(argv);
+var ae=Object.create,k=Object.defineProperty,ye=Object.getPrototypeOf,pe=Object.prototype.hasOwnProperty,de=Object.getOwnPropertyNames,he=Object.getOwnPropertyDescriptor,me=e=>k(e,"__esModule",{value:!0}),O=(e,n)=>()=>(n||(n={exports:{}},e(n.exports,n)),n.exports),be=(e,n,t)=>{if(me(e),n&&typeof n=="object"||typeof n=="function")for(let i of de(n))!pe.call(e,i)&&i!=="default"&&k(e,i,{get:()=>n[i],enumerable:!(t=he(n,i))||t.enumerable});return e},x=e=>e&&e.__esModule?e:be(k(e!=null?ae(ye(e)):{},"default",{value:e,enumerable:!0}),e),$=O((Ke,N)=>{N.exports=function(e,n){n||(n={});var t={bools:{},strings:{},unknownFn:null};typeof n.unknown=="function"&&(t.unknownFn=n.unknown),typeof n.boolean=="boolean"&&n.boolean?t.allBools=!0:[].concat(n.boolean).filter(Boolean).forEach(function(o){t.bools[o]=!0});var i={};Object.keys(n.alias||{}).forEach(function(o){i[o]=[].concat(n.alias[o]),i[o].forEach(function(d){i[d]=[o].concat(i[o].filter(function(_){return d!==_}))})}),[].concat(n.string).filter(Boolean).forEach(function(o){t.strings[o]=!0,i[o]&&(t.strings[i[o]]=!0)});var r=n.default||{},s={_:[]};Object.keys(t.bools).forEach(function(o){p(o,r[o]===void 0?!1:r[o])});var y=[];e.indexOf("--")!==-1&&(y=e.slice(e.indexOf("--")+1),e=e.slice(0,e.indexOf("--")));function m(o,d){return t.allBools&&/^--[^=]+$/.test(d)||t.strings[o]||t.bools[o]||i[o]}function p(o,d,_){if(_&&t.unknownFn&&!m(o,_)&&t.unknownFn(_)===!1)return;var u=!t.strings[o]&&B(d)?Number(d):d;S(s,o.split("."),u),(i[o]||[]).forEach(function(A){S(s,A.split("."),u)})}function S(o,d,_){for(var u=o,A=0;A<d.length-1;A++){var a=d[A];if(a==="__proto__")return;u[a]===void 0&&(u[a]={}),(u[a]===Object.prototype||u[a]===Number.prototype||u[a]===String.prototype)&&(u[a]={}),u[a]===Array.prototype&&(u[a]=[]),u=u[a]}var a=d[d.length-1];if(a==="__proto__")return;(u===Object.prototype||u===Number.prototype||u===String.prototype)&&(u={}),u===Array.prototype&&(u=[]),u[a]===void 0||t.bools[a]||typeof u[a]=="boolean"?u[a]=_:Array.isArray(u[a])?u[a].push(_):u[a]=[u[a],_]}function w(o){return i[o].some(function(d){return t.bools[d]})}for(var c=0;c<e.length;c++){var f=e[c];if(/^--.+=/.test(f)){var q=f.match(/^--([^=]+)=([\s\S]*)$/),l=q[1],g=q[2];t.bools[l]&&(g=g!=="false"),p(l,g,f)}else if(/^--no-.+/.test(f)){var l=f.match(/^--no-(.+)/)[1];p(l,!1,f)}else if(/^--.+/.test(f)){var l=f.match(/^--(.+)/)[1],b=e[c+1];b!==void 0&&!/^-/.test(b)&&!t.bools[l]&&!t.allBools&&(i[l]?!w(l):!0)?(p(l,b,f),c++):/^(true|false)$/.test(b)?(p(l,b==="true",f),c++):p(l,t.strings[l]?"":!0,f)}else if(/^-[^-]+/.test(f)){for(var v=f.slice(1,-1).split(""),F=!1,h=0;h<v.length;h++){var b=f.slice(h+2);if(b==="-"){p(v[h],b,f);continue}if(/[A-Za-z]/.test(v[h])&&/=/.test(b)){p(v[h],b.split("=")[1],f),F=!0;break}if(/[A-Za-z]/.test(v[h])&&/-?\d+(\.\d*)?(e-?\d+)?$/.test(b)){p(v[h],b,f),F=!0;break}if(v[h+1]&&v[h+1].match(/\W/)){p(v[h],f.slice(h+2),f),F=!0;break}else p(v[h],t.strings[v[h]]?"":!0,f)}var l=f.slice(-1)[0];!F&&l!=="-"&&(e[c+1]&&!/^(-|--)[^-]/.test(e[c+1])&&!t.bools[l]&&(i[l]?!w(l):!0)?(p(l,e[c+1],f),c++):e[c+1]&&/^(true|false)$/.test(e[c+1])?(p(l,e[c+1]==="true",f),c++):p(l,t.strings[l]?"":!0,f))}else if((!t.unknownFn||t.unknownFn(f)!==!1)&&s._.push(t.strings._||!B(f)?f:Number(f)),n.stopEarly){s._.push.apply(s._,e.slice(c+1));break}}return Object.keys(r).forEach(function(o){ve(s,o.split("."))||(S(s,o.split("."),r[o]),(i[o]||[]).forEach(function(d){S(s,d.split("."),r[o])}))}),n["--"]?(s["--"]=new Array,y.forEach(function(o){s["--"].push(o)})):y.forEach(function(o){s._.push(o)}),s};function ve(e,n){var t=e;n.slice(0,-1).forEach(function(r){t=t[r]||{}});var i=n[n.length-1];return i in t}function B(e){return typeof e=="number"||/^0x[0-9a-f]+$/i.test(e)?!0:/^[-+]?(?:\d+(?:\.\d*)?|\.\d+)(e[-+]?\d+)?$/.test(e)}}),K=O((Pe,I)=>{"use strict";I.exports=function e(n,t){var i=t.indexOf(".");if(!~i)return n==null?void 0:n[t];var r=t.substring(0,i),s=t.substring(i+1);return n==null?void 0:(n=n[r],s?e(n,s):n)}}),V=O((Ve,P)=>{"use strict";P.exports=function(n){return n!=null&&typeof n=="object"&&Array.isArray(n)===!1}}),Z=O((ze,z)=>{"use strict";var _e=V();function C(e){return _e(e)===!0&&Object.prototype.toString.call(e)==="[object Object]"}z.exports=function(n){var t,i;return C(n)===!1?!1:(t=n.constructor,typeof t!="function"?!1:(i=t.prototype,C(i)===!1?!1:!(i.hasOwnProperty("isPrototypeOf")===!1)))}}),W=O((Ce,j)=>{"use strict";var D=Z();function E(e,n,t,i){if(!T(e))return e;let r=i||{},s=Array.isArray(n);if(!s&&typeof n!="string")return e;let y=r.merge;y&&typeof y!="function"&&(y=Object.assign);let m=(s?n:Oe(n,r)).filter(we),p=m.length,S=e;if(!i&&m.length===1)return L(e,m[0],t,y),e;for(let w=0;w<p;w++){let c=m[w];if(T(e[c])||(e[c]={}),w===p-1){L(e,c,t,y);break}e=e[c]}return S}function L(e,n,t,i){i&&D(e[n])&&D(t)?e[n]=i({},e[n],t):e[n]=t}function Oe(e,n){let t=Se(e,n);if(E.memo[t])return E.memo[t];let i=n&&n.separator?n.separator:".",r=[],s=[];n&&typeof n.split=="function"?r=n.split(e):r=e.split(i);for(let y=0;y<r.length;y++){let m=r[y];for(;m&&m.slice(-1)==="\\"&&r[y+1]!=null;)m=m.slice(0,-1)+i+r[++y];s.push(m)}return E.memo[t]=s,s}function Se(e,n){let t=e;if(typeof n=="undefined")return t+"";let i=Object.keys(n);for(let r=0;r<i.length;r++){let s=i[r];t+=";"+s+"="+String(n[s])}return t}function we(e){return e!=="__proto__"&&e!=="constructor"&&e!=="prototype"}function T(e){return e!==null&&(typeof e=="object"||typeof e=="function")}E.memo={};j.exports=E}),M=O((Ze,G)=>{"use strict";var H=require("fs");function Ae(e,n){if(typeof n=="function"){H.readFile(e,"utf-8",function(t,i){try{i=JSON.parse(i)}catch(r){t=t||r}n(t,i)});return}return JSON.parse(H.readFileSync(e))}G.exports=Ae}),R=O((je,Q)=>{"use strict";var Ee=typeof Symbol=="function"&&typeof Symbol.iterator=="symbol"?function(e){return typeof e}:function(e){return e&&typeof Symbol=="function"&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e};function Fe(e,n){var t=0,i=[];if(Array.isArray(e))for(;t<e.length&&!(n(e[t],t,e)===!1);++t);else if((typeof e=="undefined"?"undefined":Ee(e))==="object"&&e!==null)for(i=Object.keys(e);t<i.length&&!(n(e[i[t]],i[t],e)===!1);++t);}Q.exports=Fe}),te=O((De,U)=>{"use strict";var ge=typeof Symbol=="function"&&typeof Symbol.iterator=="symbol"?function(e){return typeof e}:function(e){return e&&typeof Symbol=="function"&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},ke=function(){function e(n,t){for(var i=0;i<t.length;i++){var r=t[i];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(n,r.key,r)}}return function(n,t,i){return t&&e(n.prototype,t),i&&e(n,i),n}}();function xe(e,n){if(!(e instanceof n))throw new TypeError("Cannot call a class as a function")}var Je=K(),X=W(),Y=M(),ee=require("fs"),qe=R(),Ne=require("os"),Be=function(){function e(n,t){xe(this,e),this.options=t=t||{},t.stringify_width=t.stringify_width||2,t.stringify_fn=t.stringify_fn||null,t.stringify_eol=t.stringify_eol||!1,t.ignore_dots=t.ignore_dots||!1,this.path=n,this.data=this.read()}return ke(e,[{key:"set",value:function(t,i){var r=this;return(typeof t=="undefined"?"undefined":ge(t))==="object"?qe(t,function(s,y){X(r.data,y,s)}):this.options.ignore_dots?this.data[t]=i:X(this.data,t,i),this.options.autosave&&this.save(),this}},{key:"get",value:function(t){return t?this.options.ignore_dots?this.data[t]:Je(this.data,t):this.toObject()}},{key:"unset",value:function(t){return this.set(t,void 0)}},{key:"read",value:function(t){if(!t)try{return Y(this.path)}catch(i){return{}}Y(this.path,function(i,r){r=i?{}:r,t(null,r)})}},{key:"write",value:function(t,i){return i?ee.writeFile(this.path,t,i):ee.writeFileSync(this.path,t),this}},{key:"empty",value:function(t){return this.write("{}",t)}},{key:"save",value:function(t){var i=JSON.stringify(this.data,this.options.stringify_fn,this.options.stringify_width,this.options.stringify_eol);return this.write(this.options.stringify_eol?i+Ne.EOL:i,t),this}},{key:"toObject",value:function(){return this.data}}]),e}();U.exports=function(n,t){return new Be(n,t)}}),ue=x($()),ne=x(te()),ie=x(require("path"));function J(e){({}).toString.call(e)==="[object Object]"?console.log(JSON.stringify(e,null,2)):console.log(e)}var re=process.platform==="win32"?"ia32":"x64",oe=process.platform==="win32"?"win32":"darwin",fe=ne.default(ie.default.resolve(__dirname,"..","package.json"),{stringify_width:4});function se(e){Object.keys(e).forEach(n=>{fe.set(`agora_electron.${n}`,e[n])}),fe.save()}var le=ue.default(process.argv.slice(2));delete le._;var $e={arch:re,platform:oe,electron_version:"7.1.2",msvs_version:"2017",silent:!1,debug:!1,prebuilt:!0},ce={...$e,...le};J("current agora parameters");J(ce);se(ce);
+/*!
+ * is-plain-object <https://github.com/jonschlinkert/is-plain-object>
+ *
+ * Copyright (c) 2014-2017, Jon Schlinkert.
+ * Released under the MIT License.
+ */
+/*!
+ * isobject <https://github.com/jonschlinkert/isobject>
+ *
+ * Copyright (c) 2014-2017, Jon Schlinkert.
+ * Released under the MIT License.
+ */
+/*!
+ * set-value <https://github.com/jonschlinkert/set-value>
+ *
+ * Copyright (c) 2014-2018, Jon Schlinkert.
+ * Released under the MIT License.
+ */
