+++ scripts/package.json
@@ -1,0 +1,13 @@
+{
+    "private": true,
+    "scripts": {
+        "build": "esbuild src/index.ts --platform=node --bundle --minify --outfile=Build-agora-sdk.js"
+    },
+    "dependencies": {
+        "edit-json-file": "^1.5.0",
+        "minimist": "^1.2.5"
+    },
+    "devDependencies": {
+        "esbuild": "^0.8.23"
+    }
+}
